adding error page test and functionality

+++ src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -39,6 +39,11 @@ public class GradebookController {
 
     @GetMapping("/delete/student/{id}")
     public String deleteStudent(@PathVariable int id, Model m) {
+
+        if(!studentAndGradeService.checkIfStudentIsNull(id)){
+            return "error";
+        }
+
         studentAndGradeService.deleteStudent(id);
         m.addAttribute("students", studentAndGradeService.getGradebook());
         return "index";
